requirements.txt
paho-mqtt
python-dotenv
psycopg2
numpy
matplotlib
control
nidaqmx